app/requirements.txt
alembic>=1.5.8
anyio>=3.6.1
attrs>=20.3.0
bcrypt>=3.2.0
certifi>=2020.12.5
cffi>=1.14.5
chardet>=4.0.0
click>=7.1.2
dnspython>=2.1.0
docstring-parser>=0.7.3
email-validator>=1.1.2
Faker>=7.0.0
fastapi>=0.85.0
FastAPI-SQLAlchemy>=0.2.1
greenlet>=1.0.0
h11>=0.12.0
idna>=2.10
importlib-metadata>=5.0.0
iniconfig>=1.1.1
Mako>=1.1.4
MarkupSafe>=1.1.1
numpy>=1.21.6
packaging>=20.9
pandas>=1.3.5
passlib>=1.7.4
pluggy>=0.13.1
psycopg2-binary>=2.9.4
py>=1.10.0
pycparser>=2.20
pydantic>=1.10.2,<2.0.0
PyJWT>=2.0.1
pyparsing>=2.4.7
pytest>=6.2.2
python-dateutil>=2.8.1
python-dotenv>=0.15.0
python-editor>=1.0.4
python-multipart>=0.0.5
pytz>=2022.4
requests>=2.25.1
six>=1.15.0
sniffio>=1.3.0
SQLAlchemy>=1.4.3,<2.0.0
starlette>=0.20.4
text-unidecode>=1.3
toml>=0.10.2
typing_extensions>=4.4.0
urllib3>=1.26.4
uvicorn>=0.13.4
zipp>=3.9.0
